+++ internal/input/device.go
@@ -1,0 +1,6 @@
+package input
+
+type Device interface {
+	ReadLine() (string, error)
+	GetChar() (rune, error)
+}